server/requirements.txt
blinker
click
Flask
flask-cors
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
mysqlclient
numpy
pandas
python-dateutil
pytz
six
SQLAlchemy
typing_extensions
tzdata
Werkzeug
flask-bcrypt
PyJWT
gunicorn
scikit-learn
joblib
python-dotenv
